ALMA-37--fjernet AlmaServiceFactory.java fjernet interfaces, fjernet alma-stage (bruker string)
