+++ src/main/java/no/unit/alma/commons/AlmaClient.java
@@ -1,6 +1,8 @@
 package no.unit.alma.commons;
 
 import com.typesafe.config.Config;
+import no.bibsys.vault.AppRole;
+import no.bibsys.vault.VaultClient;
 import org.glassfish.jersey.client.ClientProperties;
 import org.glassfish.jersey.moxy.xml.MoxyXmlFeature;
 
@@ -13,15 +15,38 @@ public class AlmaClient {
 
     private final WebTarget webTarget;
     private final String contextValue;
-    private final AlmaStage almaStage;
+    private final String almaStage;
 
     private final static int connectTimeout = 10_000;
     private final static int readTimeout = 120_000;
     private final static String CONTEXT = "bibsysBibKode";
 
-    public AlmaClient(Config config, Client client, VaultApiAuthorization apiAuthorization) {
-        Objects.requireNonNull(apiAuthorization, "Alma API authorization is required");
+    public AlmaClient(Client client, Config config, String bibCode) {
+        this(client,
+                config,
+                VaultClient.builder()
+                        .withCredentials(AppRole.from(config.getString("roleId"), config.getString("secretId")))
+                        .build(),
+                bibCode
+        );
+    }
+
+    public AlmaClient(Client client, Config config, VaultClient vaultClient, String bibCode) {
+        this(client,
+                config,
+                ApiAuthorizationService.builder()
+                        .vaultClient(vaultClient)
+                        .environment(config.getString("environment"))
+                        .build(),
+                bibCode
+        );
+    }
+
+    public AlmaClient(Client client, Config config, ApiAuthorizationService apiAuthorizationService, String bibCode) {
         Objects.requireNonNull(client, "JAX-RS rest client must be provided");
+        Objects.requireNonNull(apiAuthorizationService, "Alma API authorization is required");
+        VaultApiAuthorization apiAuthorization = apiAuthorizationService.getApiAuthorization(config.getString("stage"), bibCode);
+
         this.webTarget = client
                 .property(ClientProperties.CONNECT_TIMEOUT, connectTimeout)
                 .property(ClientProperties.READ_TIMEOUT, readTimeout)
@@ -30,7 +55,7 @@ public class AlmaClient {
                 .register(new RequestResponseLogger(config.getString("app"), config.getString("stage")), 6000)
                 .register(AlmaStatusResponseFilter.class, Priorities.ENTITY_CODER)
                 .target(buildAlmaUrl(apiAuthorization.getAlmaHost(), config.getString("almaServiceContext")));
-        this.almaStage = apiAuthorization.getAlmaStage();
+        this.almaStage = config.getString("stage");
         this.contextValue = apiAuthorization.getOrganization();
     }
 
@@ -46,7 +71,7 @@ public class AlmaClient {
         return contextValue;
     }
 
-    public AlmaStage getAlmaStage() {
+    public String getAlmaStage() {
         return almaStage;
     }
