--- src/main/java/no/unit/alma/commons/AlmaContext.java
@@ -1,10 +1,0 @@
-package no.unit.alma.commons;
-
-public interface AlmaContext {
-
-    String getContext();
-    String getContextValue();
-    default AlmaStage getAlmaStage() {
-        throw new UnsupportedOperationException("Alma stage has not been implemented. Implement it if you need it!");
-    }
-}
