+++ src/main/resources/reference.conf
@@ -6,7 +6,7 @@ secretId="XXX"
 environment="dev"
 environment=${?ENVIRONMENT}
 app="test"
-stage="unknown"
+stage="alma-sandbox2"
 
 almaServiceContext="almaws/v1"
 almaServiceContext=${?ALMA_SERVICE_CONTEXT}
